Use the right field type specifier in function calls.
(/analyze warning C6340)

+++ src/mpc-hc/PlayerStatusBar.cpp
@@ -232,23 +232,23 @@ void CPlayerStatusBar::SetStatusTimer(REFERENCE_TIME rtNow, REFERENCE_TIME rtDur
         }
 
         if (tcDur.bHours > 0 || (rtNow >= rtDur && tcNow.bHours > 0)) {
-            posstr.Format(_T("%02d:%02d:%02d"), tcNow.bHours, tcNow.bMinutes, tcNow.bSeconds);
-            rstr.Format(_T("%02d:%02d:%02d"), tcRt.bHours, tcRt.bMinutes, tcRt.bSeconds);
+            posstr.Format(_T("%02u:%02u:%02u"), tcNow.bHours, tcNow.bMinutes, tcNow.bSeconds);
+            rstr.Format(_T("%02u:%02u:%02u"), tcRt.bHours, tcRt.bMinutes, tcRt.bSeconds);
         } else {
-            posstr.Format(_T("%02d:%02d"), tcNow.bMinutes, tcNow.bSeconds);
-            rstr.Format(_T("%02d:%02d"), tcRt.bMinutes, tcRt.bSeconds);
+            posstr.Format(_T("%02u:%02u"), tcNow.bMinutes, tcNow.bSeconds);
+            rstr.Format(_T("%02u:%02u"), tcRt.bMinutes, tcRt.bSeconds);
         }
 
         if (tcDur.bHours > 0) {
-            durstr.Format(_T("%02d:%02d:%02d"), tcDur.bHours, tcDur.bMinutes, tcDur.bSeconds);
+            durstr.Format(_T("%02u:%02u:%02u"), tcDur.bHours, tcDur.bMinutes, tcDur.bSeconds);
         } else {
-            durstr.Format(_T("%02d:%02d"), tcDur.bMinutes, tcDur.bSeconds);
+            durstr.Format(_T("%02u:%02u"), tcDur.bMinutes, tcDur.bSeconds);
         }
 
         if (fHighPrecision) {
-            posstr.AppendFormat(_T(".%03d"), (rtNow / 10000) % 1000);
-            durstr.AppendFormat(_T(".%03d"), (rtDur / 10000) % 1000);
-            rstr.AppendFormat(_T(".%03d"), ((rtDur - rtNow) / 10000) % 1000);
+            posstr.AppendFormat(_T(".%03d"), int((rtNow / 10000) % 1000));
+            durstr.AppendFormat(_T(".%03d"), int((rtDur / 10000) % 1000));
+            rstr.AppendFormat(_T(".%03d"), int(((rtDur - rtNow) / 10000) % 1000));
         }
     } else if (*pTimeFormat == TIME_FORMAT_FRAME) {
         posstr.Format(_T("%I64d"), rtNow);
